influx query calculations
Now calculates time for influx. Runs both 1 mont and 1 year when querying through the main CLI. Output is only logged and not saved in file.

Simple solution for getting faker data is still the same, this means that it is not possible to create data for more than 1000 devices for 1 year in this commit.

Also the executable query is with included docker access commands as running the query through the API would have required the Node.js program to be executed inside of the database container for the influx API to work without modifying it.